[misc/spotify] Move json helper functions into its own file
